requirements.txt
hazm==0.10.0
langchain==0.1.6
langdetect==1.0.9
nltk==3.8.1
pinecone_client==2.2.2
PyPDF2==3.0.1
python_docx==1.1.0
sentence_transformers==2.2.2
torch==2.0.1
transformers==4.31.0
